refactor: dependency inject quest db
This gives us a path to test out other implementations or to switch what
we test with for the quest database. Also cleaned up one of the other
dependency injected modules whose name was wrong.

I put the injector for the repo in the quests module instead of
PointQuest since this is a private detail to that context and not
something others should need to look up.

Closes: PD-27