feat: get deployed on fly.io
Needed to remove all of the SQL references since we do not have a
database at the moment.

Fly config was generated by running

```bash
fly launch
```

And then removing the migration scripts that it generated.

To deploy a second time we can now do

```bash
fly deploy --ha=false
```

Without the `ha` flag two machines gets deployed which won't work with
the game state until we get that distributed.