refactor: persist form state on change
Adding a phx-change handler to the form allows us to not lose data when switching
an element that is in the form but outside of the changeset. We just need to
update the form state and put it back on the socket so that it is persistent
when LiveView re-renders the form.