fix: prevent changeset errors from crashing form
It is normal and valid for the changeset to load up in an
error state when first hitting the quest start page. This ensures
that errors on the changeset at launch do not crash the liveview.

Also removed dead code as the command on the socket was never being used.