refactor: only return events from commands
There was a few places where we were returning the quest state and the
event that occurred from executing a command. Doing this means that we
are going to make it harder to change our internal model if we need to.
As much as possible we should bind to events and keep them as minimal as
we can.

Closes: PQ-80