feat: add start quest telemetry event and log handler
Start quest was one of the commands that was following the old pattern of
returning quest state instead of the quest event. Went ahead and updated
that and had to temporarily call into the db to get the quest state. Up
next I will write a command to get quest by id.

Also the `iex.exs` file was going to get really messy with dispatching
the event and then querying the database for each quest. Since we aren't
using it that often I went a head and removed all the seeder like calls
we were making in there.

Closes: PQ-33