feat: visually indicate when someone has attacked
Now that we have attacks fully evented and the liveview subscribing to these events,
we need to be able to indicate to connected parties that a given adventurer has
attacked. This will allow the Party Leader to know once all adventurers are attacking,
and they can then reveal the attack values (forthcoming).