feat: capture quest objective when starting new round
In lieu of having an integration to an issue system like Linear, we allow
the party leader to provide a link to the current ticket. This will update
in their view as well as the adventurer's view so that everyone has a link
to the current focus of the round, and knows what we're attacking.

In the future, we would like to update this logic to dynamically dispatch
to the issue provider (Linear, Github, Jira, etc.) in order to populate
the issue view into the quest state to be shared amongst all the
adventurers without needing to leave the session.