feat: fully implement projectionist reader
Up until this point we haven't really been using projection patterns
since we were only applying one event at a time on top of a mutable
model held in the QuestServer. This meant that if events came in at the
correct time they could blow each others state away.

```
bob quest -> attack -> write(bob_quest, attack)
alice quest -> attack -> write(alice_quest, attack)
```

With this new change we do not mutate the model directly and instead
buffer events into a log. Once a certain count of events is hit we drop
a few and generate a snapshot. This should work well given writers
append to the end and the snapshot drops from the front. Also the quest
server is a single threaded queue so while doing the snapshotting it
will not be performing other tasks.

```
bob quest -> attack -> write(bob_quest, attack) -> events ++ [attack]
alice quest -> attack -> write(alice_quest, attack) -> events ++ [attack]
```

So now even if you have a stale quest when you go to write it is not an
issue since that event will be appended to the log and the quest past in
isn't used to derive anything. A new reprojection happens with all
events on top of previous snapshot. Hopefully no race conditions!

Lastly, there were a lot of weird behaviors happening around leader and
leader adventurer ids. By dropping them from the command and generating
them in the projection it was a lot easier to line up wires and make it
so that ids were consistent. Also makes sense to have the same id
between leader and their optional adventurer

Closes: PQ-92