refactor: ui polish
Lots of UI changes:
  * Turned the attack tiles into diamonds
  * Centered attack diamonds
  * Added attack tile
  * Customized partly leader toolbar
  * Add some icons

On the party leader toolbar I wanted to try and optimize the controls
for what the party leader would be engaging with the most. The invite
link makes sense to have on the far right since it isn't interacted with
as much as the rounds are being driven. Hopefully moving the core
controls to the right makes things easier to work with.

Closes: PQ-98