feat: ensure actor is not nil before loading quest page
When presented with a `nil` actor, the /quest/[:id] route blows up in
fun and exciting ways. Adding the EnsureActor hook enforces that if
a user makes it to this route, there is a usable actor on the socket.

If a nil actor is found in this hook, we instead redirect them to the
quest start page.