hotfix: starting a quest
We had more quest id confusion that was causing some paths to generate a
different quest id that would fail further lookups. I think this is a
good lesson that we should avoid auto id generation in as many spots as
possible and handle it up front in commands and events.