feat: add sound effects pipeline
I went the route of implementing this mostly client side. By leveraging
localStorage we don't have to store this in a database and have it be
expensive to toggle on or off. It also makes the logic pretty
lightweight on the backend. For initial effects there is a click noise
on attack and a win sound when all votes are the same.

**Backend**

On the backend I decided to add a new nested liveview which is also
subscribed to pubsub for quest events. This makes it so that the core
quest liveview doesn't also have to juggle sounds and can just focus on
state changes. The live view for audio will always push sound events and
let the client decide if it should play them or not.

**Frontend**

On the frontend I went with HTML custom elements. This made it nice to
define the component logic without stooping to too low level of JS. I
also really like from the backend how the `<audio-preferences>`
component stands out and you have a strong signal where custom js
exists.

Custom Elements have a "feature" of isolating styles of the component
from the outside context. The logic to create a custom link element in
the component was so that I could pull in the outer style sheet. With
browser caching this shouldn't load the stylesheet again. This hint was
given by Mozilla in their docs:
https://developer.mozilla.org/en-US/docs/Web/API/Web_components/Using_custom_elements#referencing_external_styles

To keep button styles in sync I moved button styling to tailwind so that
both phoenix and js are in line.

Closes: PQ-107