refactor: move current db to in memory namespace
With the couch implementation incoming its nice to get the two
implementations separated from each other so that they can be toggled if
needed. To make the implementation as self contained as possible I even
included a supervisor that bundles that part of the tree. This way we
can toggle off the whole supervisor in the case where we don't want the
inmemory in production anymore.

Closes: PQ-128