remove adventurer from party UI
Handles the emitted event on the front end to safely remove the
adventurer (from the perspective of the rest of the party). Currently,
this puts the removed adventurer in a bad state where they would need to
clear their cookies to remove the phoenix session token. This also does
not allow a Party Leader to remove their own adventurer.

Additionally, the UI to remove a member is also not included in this PR.