feat: add ui for project manager to remove a member
Move the current controls (previously only alerting an adventurer) to
only show on hover, and add a control to remove a member from the quest.

Known issues:

- Party leader's cannot remove themselves. Removing your adventurer as
the party leader will remove them from the UI, but only until a refresh
or the quest is re-queried, as the adventurer is still parsed from their
party leader object and not the adventurer's list.

- Adventurer's that are removed end on an error page as we're not
removing their session tokens currently.