refactor: return event with id from db write
In order to make snapshots work for projecting events, we need to ensure
that the events are strongly orderable.

All events should now have a primary key of id.

Valuable changed to ignore the id key on event changesets.

Commands now return {:ok, event} directly from the database call to
write the event.