feat: implement quest snapshots
Utilize couch's start_key and descending sorting to allow retrieving of
snapshots including finding the most recent snapshot before an event.
Add some tests thoroughly proving the logic in this convoluted case
statement.

Extend some of the command functions for testability to allow injecting
the underlying repo dependency to allow for testing Couch via these unit
tests when the other tests are utilizing the in-memory DB for unit
testing.