feat: automatically snapshot after set number of events\
Re-integrated Horde to enforce the quest server being a singleton
implementation. Refactored the quest server to automatically check if it
has surpassed the threshold for the required number of events to create
a snapshot, and if so, automatically schedules a snapshot to be taken at
1/2 of the current allowed number of events. For instance, with a window
of 50 events, a snapshot would be scheduled on receiving the 50th event,
causing event 25 to be the new snapshot location.